# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PakKepo
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://pak_kepo.test'
paths:
...
  /api/login:
    post:
      summary: Authentication
      operationId: authentication
      description: 'Handle an incoming authentication request.'
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: oconner.angelo@example.com
                password:
                  type: string
                  description: ''
                  example: 'b^I,~n'
              required:
                - email
                - password
      security: []
...